refactored w/state in app.js
