+++ src/Form.js
@@ -2,132 +2,26 @@ import React from 'react';
 import './Form.css';
 
 class Form extends React.Component {
-  constructor() {
-    super();
-    this.state = {
-      numbers: [],
-      result: '',
-      operation: '',
-      input: '',
-    };
-  }
-
-  validInputChecker = () => {
-    const { numbers, operation } = this.state;
-    if (numbers.length === 0) {
-      return false;
-    } else if (
-      numbers.some((num) => {
-        return Number.isNaN(num);
-      })
-    ) {
-      return false;
-    } else if (
-      operation !== 'sum' &&
-      operation !== 'average' &&
-      operation !== 'mode'
-    ) {
-      return false;
-    }
-    return true;
-  };
-
-  getSum = (arr) => {
-    let sum = 0;
-    for (let num of arr) {
-      sum += num;
-    }
-    return sum;
-  };
-
-  // make a countObj to keep track of how many of each num exists, then make an array to find the key with hightest value
-  getMode = (arr) => {
-    let countObj = {};
-    for (let num of arr) {
-      if (countObj[num]) {
-        countObj[num]++;
-      } else {
-        countObj[num] = 1;
-      }
-    }
-
-    // // at this point, countObj is made up of the digits of user input as keys and the # of times they appear as their values.
-    let keys = Object.keys(countObj);
-    let highestValue = 0;
-    let highestKey;
-    for(let key of keys){
-      if(countObj[key] > highestValue){
-        highestValue = countObj[key];
-        highestKey = key;
-      }
-    }
-    return highestKey
-  };
-
-  operationHandler = (event) => {
-    const { value } = event.target;
-    this.setState({
-      operation: value,
-    });
-  };
-
-  inputHandler = (event) => {
-    const { value } = event.target;
-    // take apart the numbers inputted and put them back into an array in #form
-    const arr = value.split(',').map((num) => Number(num));
-    this.setState({
-      numbers: arr,
-      input: value,
-    });
-  };
-
-  formSubmitHandler = (event) => {
-    event.preventDefault();
-    const { operation, numbers } = this.state;
-
-    if (!this.validInputChecker()) {
-      this.setState({
-        result: 'Invalid input.',
-        input: '',
-        operation: '',
-      });
-    } else {
-      if (operation === 'sum') {
-        this.setState({
-          result: this.getSum(numbers),
-          input: '',
-          operation: '',
-        });
-      } else if (operation === 'average') {
-        this.setState({
-          result: this.getSum(numbers) / numbers.length,
-          input: '',
-          operation: '',
-        });
-      } else {
-        this.setState({
-          result: this.getMode(numbers),
-          input: '',
-          operation: '',
-        });
-      }
-    }
-  };
-
   render() {
-    const { result, input, operation } = this.state;
+    const {
+      input,
+      operation,
+      inputHandler,
+      operationHandler,
+      formSubmitHandler,
+    } = this.props;
     return (
       <>
-        <form onSubmit={this.formSubmitHandler}>
+        <form onSubmit={formSubmitHandler}>
           <input
-            onChange={this.inputHandler}
+            onChange={inputHandler}
             id="values"
             name="values"
             type="text"
             value={input}
           />
           <select
-            onChange={this.operationHandler}
+            onChange={operationHandler}
             id="operation"
             name="operation"
             value={operation}
@@ -139,7 +33,6 @@ class Form extends React.Component {
           </select>
           <button type="submit">Calculate</button>
         </form>
-        <p>{result}</p>
       </>
     );
   }
